Move addKeyDownHandler function to separate file
