+++ src/lib/renderComponent.tsx
@@ -8,7 +8,6 @@ import getClasses from './getClasses'
 import getElementType from './getElementType'
 import getUnhandledProps from './getUnhandledProps'
 import toCompactArray from './toCompactArray'
-import * as _ from 'lodash'
 import {
   ComponentStyleFunctionParam,
   ComponentVariablesInput,
@@ -22,7 +21,7 @@ import {
 } from '../../types/theme'
 import { IAccessibilityDefinition, AccessibilityActions } from './accessibility/interfaces'
 import { DefaultBehavior } from './accessibility'
-import keyboardHandlerFilter from './accessibility/Helpers/keyboardHandlerFilter'
+import addKeyDownHandler from './addKeyDownHandler'
 import { mergeComponentStyles, mergeComponentVariables } from './mergeThemes'
 
 export interface IRenderResultConfig<P> {
@@ -58,42 +57,6 @@ const getAccessibility = <P extends {}>(props, state) => {
     ...props,
     ...state,
   })
-}
-
-/**
- * Adds onKeyDown handler to the Component's rest props, based on Component's actions
- * and keys mappings defined in Accessibility behavior
- * @param {Object} rest The rest props which is to be extended by adding onKeyDown handler
- * @param {AccessibilityActions} actions The input element which is to loose focus.
- * @param {IAccessibilityDefinition} accessibility The input element which is to loose focus.
- * @param {IRenderConfigProps} props The props which are used to invoke onKeyDown handler passed from top.
- */
-const addKeyDownHandler = (rest, actions, accessibility, props) => {
-  const actionsDefinition = accessibility.actionsDefinition
-
-  if (!actions || !actionsDefinition) return
-
-  let hasCommonActions = false
-  for (const actionName in actionsDefinition) {
-    if (actions[actionName]) {
-      hasCommonActions = true
-      break
-    }
-  }
-  if (!hasCommonActions) return
-
-  rest.onKeyDown = (event: React.KeyboardEvent) => {
-    for (const actionName in actionsDefinition) {
-      if (!actions[actionName]) continue
-      const eventHandler = keyboardHandlerFilter(
-        actions[actionName],
-        actionsDefinition[actionName].keyCombinations,
-      )
-      eventHandler && eventHandler(event)
-    }
-
-    _.invoke(props, 'onKeyDown', event)
-  }
 }
 
 const renderComponent = <P extends {}>(
